fix: 삭제 확인 창 수정

See merge request s11-webmobile1-sub2/S11P12C205!140

+++ front/ott/src/components/lookbook/LookbookDetail.jsx
@@ -102,15 +102,11 @@ const LookbookDetail = ({
   // });
 
   const handleEditLookbook = () => {
-    const isConfirmed = window.confirm('룩북을 삭제하시겠습니까?');
-
-    if (isConfirmed) {
-      hideDetail();
-      // console.log(lookbook, '룩북!!');
-      nav(`/update-lookbook/${lookbookId}`, {
-        state: { lookbook: lookbook },
-      });
-    }
+    hideDetail();
+    // console.log(lookbook, '룩북!!');
+    nav(`/update-lookbook/${lookbookId}`, {
+      state: { lookbook: lookbook },
+    });
   };
 
   if (!lookbook) {
@@ -190,6 +186,12 @@ const LookbookDetail = ({
   };
 
   const handleDelete = async () => {
+    const isConfirmed = window.confirm('정말로 삭제하시겠습니까?');
+
+    if (!isConfirmed) {
+      return; // 사용자가 취소를 누르면 함수가 여기서 종료됩니다.
+    }
+
     try {
       await lookbookDelete(lookbookId);
       deleteLookbook(lookbookId);
